requirements.txt
#python=3.10
pandas=1.4.1
boto3==1.21.7
requests==2.27.1

# For DEV
python-dotenv==0.19.2
# pre-commit==2.17.0
# flake8==4.0.1
